add auth to delete task

+++ src/routers/task.js
@@ -96,11 +96,11 @@ router.patch('/tasks/:id', auth, async (req, res) => {
     }
 })
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id
 
     try {
-        const task = await Task.findByIdAndDelete(_id)
+        const task = await Task.findByIdAndDelete({_id, owner: req.user._id})
 
         if (!task) {
             res.status(400).send()
